backend/requirements.txt
# Backend Requirements
fastapi==0.128.0
uvicorn==0.40.0
python-docx==1.2.0
python-multipart==0.0.6
